backend/requirements.txt
# =====================================================
# CORE FRAMEWORK
# =====================================================
fastapi==0.110.0
uvicorn[standard]==0.27.1

# =====================================================
# DATABASE & ORM (SYNC + FUTURE ASYNC READY)
# =====================================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9        # PostgreSQL (sync)
asyncpg==0.29.0               # PostgreSQL (async, future use)

# =====================================================
# DATA VALIDATION & CONFIG
# =====================================================
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# =====================================================
# AUTHENTICATION & SECURITY
# =====================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-users[sqlalchemy]==12.1.2

# =====================================================
# FILE UPLOAD & BACKGROUND TASKS
# =====================================================
python-multipart>=0.0.6,<0.0.7
aiofiles==23.2.1

# =====================================================
# DATABASE MIGRATIONS
# =====================================================
alembic==1.13.1

# =====================================================
# HTTP CLIENT (INTERNAL / TESTING / INTEGRATION)
# =====================================================
httpx==0.26.0
requests==2.31.0

# =====================================================
# CACHING & PERFORMANCE (OPTIONAL FUTURE)
# =====================================================
redis==5.0.1

# =====================================================
# LOGGING & MONITORING
# =====================================================
loguru==0.7.2

# =====================================================
# EMAIL / NOTIFICATIONS (MONTH 4+ READY)
# =====================================================
emails==0.6
jinja2==3.1.3

# =====================================================
# TESTING (OPTIONAL BUT READY)
# =====================================================
pytest==8.0.2
pytest-cov==4.1.0

# =====================================================
# SECURITY & RATE LIMITING (OPTIONAL)
# =====================================================
slowapi==0.1.9

# =====================================================
# UTILITIES
# =====================================================
tenacity==8.2.3
